Small fixes

+++ birdset/modules/base_module.py
@@ -199,7 +199,7 @@ class BaseModule(L.LightningModule):
             prog_bar=True
         )
 
-        self.valid_metric(preds, targets)
+        self.valid_metric(preds, targets.int())
         self.log(
             f"val/{self.valid_metric.__class__.__name__}",
             self.valid_metric,
